check that blackbox work with a fix
prior to a fix, unhandled error related
to export token was issued by browser

moved blackbox to a separate module

this works if clicking GBAD: Export as
RDF/XML

bun run test; bun run build; bun run serve,
all ok
